+++ src/main/scala/io/kaizensolutions/virgil/PullMode.scala
@@ -3,6 +3,6 @@ package io.kaizensolutions.virgil
 sealed trait PullMode
 object PullMode {
   type All = All.type
-  case object All             extends PullMode
-  case class TakeUpto(n: Int) extends PullMode
+  case object All              extends PullMode
+  case class TakeUpto(n: Long) extends PullMode
 }
